eeg_attention_decoder/requirements.txt
# EEG Attention Decoder — Python Dependencies
# Install with: pip install -r requirements.txt
# Python >= 3.11 required

# Core scientific stack
numpy>=1.26.0
scipy>=1.12.0
pandas>=2.2.0

# Machine learning
scikit-learn>=1.4.0
joblib>=1.3.0

# Configuration
PyYAML>=6.0.1

# Serial communication
pyserial>=3.5

# Realtime visualization (optional but needed for 'decode' mode)
pyqtgraph>=0.13.4
PyQt6>=6.6.0          # or PyQt5; pyqtgraph supports both

# Plotting (offline reports)
matplotlib>=3.8.0

# Testing
pytest>=8.0.0

# Reproducibility
gitpython>=3.1.40
